pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "alembic>=1.17.1",
    "authlib>=1.6.5",
    "discord-py[voice]>=2.6.4",
    "fastapi>=0.121.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "itsdangerous>=2.2.0",
    "psutil>=7.1.3",
    "psycopg2-binary>=2.9.11",
    "pynacl>=1.5.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
